+++ src/app/(editor)/project/page.tsx
@@ -5,14 +5,13 @@ import { useState } from 'react';
 import { useFieldArray, useFormContext } from 'react-hook-form';
 import { v4 } from 'uuid';
 
+import { Button, Guide, Input, Label } from '@/components/common';
 import {
-  Button,
-  Guide,
-  Input,
-  Label,
+  AiSuggestion,
+  FormCard,
   MarkdownInput,
-} from '@/components/common';
-import { AiSuggestion, FormCard, PeriodInput } from '@/components/Form';
+  PeriodInput,
+} from '@/components/Form';
 import FormRemoveButton from '@/components/Form/FormRemoveButton';
 import IconChatGpt from '@/components/Icon/IconChatGpt';
 import { INITIAL_VALUE, PLACEHOLDER } from '@/constants/form';
