+++ src/components/Form/index.ts
@@ -1,3 +1,4 @@
 export { default as AiSuggestion } from './AiSuggestion';
 export { default as FormCard } from './FormCard';
+export { default as MarkdownInput } from './MarkdownInput';
 export { default as PeriodInput } from './PeriodInput';
