+++ src/components/common/index.ts
@@ -4,7 +4,6 @@ export { default as FormErrorMessage } from './FormErrorMessage';
 export { default as Guide } from './Guide';
 export { default as Input } from './Input';
 export { default as Label } from './Label';
-export { default as MarkdownInput } from './MarkdownInput';
 export { default as Progress } from './Progress';
 export { default as ServiceErrorMessage } from './ServiceErrorMessage';
 export { default as Textarea } from './Textarea';
